Move record type definitions into a schemas/ package
Relocate the loose top-level records.py and record_types.py into a
dedicated schemas/ package so shared data-shape definitions have a home
and aren't scattered at the package root:

- schemas/base.py      <- records.py      (RecordType contract)
- schemas/regulations.py <- record_types.py (concrete RECORD_TYPES)
- schemas/__init__.py re-exports RecordType + the regulations registry,
  mirroring the sources/ and pipelines/ base+concrete convention.

Future record/type definitions for other domains can live alongside
these. All consumers updated to import from spicy_regs.schemas.
