+++ src/spicy_regs/schemas/base.py
